technical-indicator-platform/backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
yfinance==0.2.18
pandas==2.0.0
numpy==1.24.0
ta==0.10.2
python-dateutil==2.8.2
psycopg2-binary==2.9.6